LayoutTests/fast/css/outline-auto-zoom-canvas-expected.txt

This test ensures that the canvas drawFocusIfNeeded focus ring scales with CSS zoom and canvas scale.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS pixels2xZoom > pixels1x is true
PASS pixels2xScale > pixels1x is true
PASS successfullyParsed is true

TEST COMPLETE
